pyproject.toml
[project]
name = "ytcapture"
version = "0.1.0"
description = "Extract video frames and transcripts from YouTube videos into Obsidian-compatible markdown"
authors = [
    { name = "Joe Monaco", email = "joe@selfmotion.net" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["youtube", "transcript", "frames", "obsidian", "markdown", "video"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "yt-dlp>=2024.0.0",
    "youtube-transcript-api>=0.6.0",
    "ffmpeg-python>=0.2.0",
    "click>=8.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0.0",
    "imagehash>=4.3.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
ytcapture = "ytcapture.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/ytcapture"
Repository = "https://github.com/yourusername/ytcapture"
Issues = "https://github.com/yourusername/ytcapture/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ytcapture"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=ytcapture",
    "--cov-report=term-missing",
    "--cov-report=html",
    "-v"
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "yt_dlp.*",
    "youtube_transcript_api.*",
    "ffmpeg.*",
    "imagehash.*",
]
ignore_missing_imports = true
